tidy(ui): move launchpad components to ui dir
