+++ invokeai/frontend/web/src/features/ui/layouts/CanvasLaunchpadPanel.tsx
