+++ invokeai/frontend/web/src/features/ui/layouts/InitialStateMainModelPicker.tsx
