+++ invokeai/frontend/web/src/features/ui/layouts/LaunchpadUseALayoutImageButton.tsx
