+++ invokeai/frontend/web/src/features/ui/layouts/UpscalingLaunchpadPanel.tsx
