+++ invokeai/frontend/web/src/features/ui/layouts/WorkflowsLaunchpadPanel.tsx
